bot/requirements.txt
aiogram==3.10.0
openai==1.40.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.2